Removed interface UserRoleDAO
